code refactor - moved auditService and facilityService files to services folder
