+++ src/client/services/clinicPortalServices/facilityServices.js
@@ -1,5 +1,5 @@
-import { serviceConstants } from "../patientPortalServices/constants";
-import { getUserAuthToken,getUserRole,getUserID } from "../utils/util";
+import { serviceConstants } from "../../patientPortalServices/constants";
+import { getUserAuthToken,getUserRole,getUserID } from "../../utils/util";
 
 export const fetchFacilitiesForOrders = () => {	
 	var token  = getUserAuthToken();
